Sonarqube community 7.9.5 doesn't work
I tried to install with that version but isn't work.
It looks Sonarqube changed their internal policies, no longer allows run as especific user ().
The new version LTS (8.9) allows it, I think that this change isn't intrusive to project and to people like me that who are doing the course.